% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypeRimg.R
\name{dimRed}
\alias{dimRed}
\title{Dimension reduction}
\usage{
dimRed(obj, type = "tSNE")
}
\arguments{
\item{obj}{hypeRimg object}

\item{type}{type, can be umap or tSNE (default). The latter
uses FI-tSNE (https://github.com/KlugerLab/FIt-SNE).}
}
\description{
Performs dimension reduction on preselected
features
}
