% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypeRimg.R
\docType{class}
\name{hypeRimg-class}
\alias{hypeRimg-class}
\alias{hypeRimg}
\title{A S4 class representing a single image}
\description{
A S4 class representing a single image
}
\section{Slots}{

\describe{
\item{\code{cellprofiler}}{}
}}

