% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypeRimg.R
\name{loadCellP}
\alias{loadCellP}
\title{Load cellprofiler data}
\usage{
loadCellP(obj, file)
}
\arguments{
\item{file}{filename}
}
\description{
Load cell profiler output for cell segmentation and
intensity features per cell, usually found in cells.txt
Only features starting with Intensity_ are retained, for visualization
also the features (columns) starting with Location_ are needed.
}
