% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypeRimg.R
\name{visualize}
\alias{visualize}
\title{Visualize clusters}
\usage{
visualize(obj, ...)
}
\arguments{
\item{obj}{hypeRimg instance}

\item{split}{can be T (default) or F}

\item{off}{size for visualization of the clusters}

\item{rgb}{color for background image}
}
\description{
Show previously identified clusters in a representative
image
}
